Update .gitignore, changed Python logging, added exception handling.  (#62)
* Update .gitignore

Added a number of lines to .gitignore to make the lives of people who develop in PyCharm a bit more pleasant. pyCharm adds some virtual env directories and .idea directories which you do not want to end up at github.com

* Update generate_databases.py

Changed all standard print commands to Python logging commands to improve logging. Script will still print to screen, however now in a proper formatted manner and it provides an easy way start to include logging to file by adding a file handler to the logger logic.

* Update download_bigslice_hmmdb

Removed print as the main method to print to the screen and replaced this with a proper Python logging option. Added exception handling for cases where urllib.request.urlretrieve fails in retrieving the file from the external source.