Bugfix: Explicitly use "amino" alphabet when reading in CDS multifasta (#86)
* Bugfix: Explicitly use "amino" alphabet when reading in CDS multifasta

This PR fixes #82, which can be caused by the `pyhmmer.easel` module not being able to guess the alphabet of a `SequenceFile` (i.e. `easel.dna()`, `easel.rna()`, or `easel.amino()` from the first sequence alone.

* Explicitly use easel.Alphabet.amino() in all pyhmmer scans

Fixes additional cases of pyhmmer not being able to guess the alphabet automatically, which often happens when the first sequence record is shorter than 10 characters.